chore(is-op): Move the integrated service operator workflow code 'closer' to integrated services
